working on bundling the handler
