+++ packages/nextjs-lambda-origin-request/src/global.d.ts
@@ -1,4 +1,4 @@
-import { NextLambdaOriginRequestManifest } from "./types";
+import type { NextLambdaOriginRequestManifest } from "./index";
 
 declare module "*.json" {
   const value: NextLambdaOriginRequestManifest;
