+++ packages/nextjs-lambda-origin-request/webpack.config.js
@@ -1,0 +1,24 @@
+const path = require("path");
+
+module.exports = {
+  entry: "./src/index.ts",
+  module: {
+    rules: [
+      {
+        test: /\.tsx?$/,
+        use: "ts-loader",
+        exclude: /node_modules/
+      }
+    ]
+  },
+  resolve: {
+    extensions: [".tsx", ".ts", ".js"]
+  },
+  output: {
+    filename: "bundle.js",
+    path: path.resolve(__dirname, "dist")
+  },
+  externals: {
+    "./manifest.json": "./manifest.json"
+  }
+};
